Rename log_entry to log_message for better clarity and consistency - part 2
