+++ src/core/assistant-message/__tests__/log-message.test.ts
