+++ src/core/prompts/sections/index.ts
@@ -3,7 +3,7 @@ export { getSystemInfoSection } from "./system-info"
 export { getObjectiveSection } from "./objective"
 export { addCustomInstructions } from "./custom-instructions"
 export { getSharedToolUseSection } from "./tool-use"
-export { getLogEntrySection } from "./log-entry"
+export { getLogMessageSection } from "./log-message"
 export { getMcpServersSection } from "./mcp-servers"
 export { getToolUseGuidelinesSection } from "./tool-use-guidelines"
 export { getCapabilitiesSection } from "./capabilities"
