+++ src/schemas/log-message.ts
@@ -3,9 +3,9 @@ import { z } from "zod"
 export const logLevels = ["debug", "info", "warn", "error"] as const
 export const logLevelsSchema = z.enum(logLevels).optional().default("info")
 
-export const logEntryParamsSchema = z.object({
+export const logMessageParamsSchema = z.object({
 	message: z.string(),
 	level: logLevelsSchema,
 })
 
-export type LogEntryParams = z.infer<typeof logEntryParamsSchema>
+export type LogMessageParams = z.infer<typeof logMessageParamsSchema>
